+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.4
 
 ## v0.4.14
+  - Deprecate `bias` in favor of `use_bias` for `RNNCell`.
   - Add `use_bias` kwarg to `LSTMCell` and `GRUCell`
 
 ## v0.4.12